+++ tests/scheduler/backends/qblox/test_helpers_qblox.py
@@ -32,6 +32,26 @@ def test_get_nco_phase_arguments(phase, expected_steps):
     assert helpers.get_nco_phase_arguments(phase) == expected_steps
 
 
+@pytest.mark.parametrize(
+    "frequency, expected_steps",
+    [
+        (-500e6, -2000000000),
+        (-200e3, -800000),
+        (0.0, 0),
+        (200e3, 800000),
+        (500e6, 2000000000),
+    ],
+)
+def test_get_nco_set_frequency_arguments(frequency: float, expected_steps: int):
+    assert helpers.get_nco_set_frequency_arguments(frequency) == expected_steps
+
+
+@pytest.mark.parametrize("frequency", [-500e6 - 1, 500e6 + 1])
+def test_invalid_get_nco_set_frequency_arguments(frequency: float):
+    with pytest.raises(ValueError):
+        helpers.get_nco_set_frequency_arguments(frequency)
+
+
 def __get_frequencies(
     clock_freq, lo_freq, interm_freq, downconverter_freq, mix_lo
 ) -> helpers.Frequencies: