+++ CHANGELOG.md
@@ -15,7 +15,9 @@
 
 - Zhinst backend - Added distortion corrections (!600)
 - Compilation - Add `determine_relative_latencies` that determines latencies for all port-clock combinations in the hardware config relative to the minimum latency (!566, #379)
+- Operations - Introduce `SetClockFrequency` operation (!575, follow-up to !539 !543)
 - Qblox backend - Introduce `qblox.helpers.determine_clock_lo_interm_freqs` and use in `QbloxBasebandModule.assign_frequencies` and `QbloxRFModule.assign_frequencies` (!574) 
+- Qblox backend - Compile `SetClockFrequency` operation to `set_freq` + `upd_param` of 8 ns (!575, follow-up to !539 !543)
 - Documentation - Qblox backend reference guide overhaul and fix docs generation warnings (!587)
 - JSON utilities - Introduce serialization of python objects to a dotted import string, next to the already existing deserialization (`import_python_object_from_string`). Serialization now happens automatically in `DataStructure`s, deserialization is implemented using Pydantic validators in the configuration models.